+++ Package.swift
@@ -14,9 +14,9 @@ let package = Package(
         .package(url: "https://github.com/swiftlang/swift-docc-plugin.git", from: "1.4.0"),
         .package(url: "https://github.com/bare-swift/swift-bytes.git", from: "0.1.0"),
         .package(url: "https://github.com/bare-swift/swift-deflate.git", from: "0.2.0"),
-        .package(url: "https://github.com/bare-swift/swift-gzip.git", from: "0.3.0"),
-        .package(url: "https://github.com/bare-swift/swift-zlib.git", from: "0.3.0"),
-        .package(url: "https://github.com/bare-swift/swift-brotli.git", from: "0.3.0")
+        .package(url: "https://github.com/bare-swift/swift-gzip.git", from: "0.4.0"),
+        .package(url: "https://github.com/bare-swift/swift-zlib.git", from: "0.4.0"),
+        .package(url: "https://github.com/bare-swift/swift-brotli.git", from: "0.4.0")
     ],
     targets: [
         .target(